Fix unsigned vs. signed clash

+++ ssl/ssltest.c
@@ -991,7 +991,7 @@ int doit_biopair(SSL *s_ssl, SSL *c_ssl, long count,
 			do
 				{
 				size_t num;
-				int r;
+				size_t r;
 
 				r1 = BIO_ctrl_pending(io1);
 				r2 = BIO_ctrl_get_write_guarantee(io2);
@@ -1055,7 +1055,7 @@ int doit_biopair(SSL *s_ssl, SSL *c_ssl, long count,
 					
 					r = BIO_nwrite0(io1, &dataptr);
 					assert(r > 0);
-					if (r < num)
+					if (r < (int)num)
 						num = r;
 					r = BIO_read(io2, dataptr, (int)num);
 					if (r != (int)num) /* can't happen */
